Move function into the only file that uses it
